- Removed the dependency on the "impossible to install in ubuntu" SDL_console-2.1. I wrote a custom console that has a very similar API. There is a "base class" console that handles the basic IO (egoboo_console_t), and then there is a "derived class" that handles the lua integration (lua_console_t). If someone wanted to use python or tcl or pascal or lisp or cobol or whatver, you would just have to make a command handler "class" in the same way that I constructed the lua_console_t...
- added a new fixed point font for the console

- added scripts use_swig_win32.bat and use_swig_nix.sh that semi-automate the use of swig to generate the file ego_wrap.c for the lua extension. The .bat file will need to be edited to point to your actual swig executable

+++ game/graphic_fan.c
@@ -21,8 +21,8 @@
  * World mesh drawing.
  */
 
-#include "egoboo.h"
 #include "graphic.h"
+#include "egoboo.h"
 
 //--------------------------------------------------------------------------------------------
 //--------------------------------------------------------------------------------------------
@@ -46,7 +46,7 @@ void render_fan( Uint32 fan )
 
     pmem  = &(mesh.mem);
 
-    if( INVALID_TILE == fan || fan >= mesh.info.tiles_count ) return;
+    if ( INVALID_TILE == fan || fan >= mesh.info.tiles_count ) return;
     ptile = pmem->tile_list + fan;
 
     // vertex is a value from 0-15, for the meshcommandref/u/v variables
@@ -116,7 +116,7 @@ void render_fan( Uint32 fan )
         if ( meshnotexture )
         {
             GLTexture_Bind( NULL );
-            meshlasttexture = ~0;
+            meshlasttexture = (Uint16)(~0);
         }
         else
         {
